TradePlacer fixing updating/flow errors
The drawing tool is based on the risk reward one. This base drawing tool also has some design problems, being the main one not updating all charts when data is updated in one of them. This was fixed by always pulling the price from the anchors prior to generating the price text string. The base drawing tool also always forces a fixed RR and that was removed leading to a small redesign in the setReward/setRisk -> updateBaseData